feat: huge number of changes to get the libs all compiling in 64bits. Some things may be utterly broken. Some linking issues at the end of compilation with ICCItemPool, inteljpegutil and zlib

+++ Alchemy/CodeChain/CTextFilePage.cpp
@@ -69,7 +69,7 @@ ICCItem *CTextFilePage::EnumGetNext (CCodeChain &CC, SPageEnumCtx &Ctx)
 
 	//	Set context
 
-	Ctx.dwPos1 = pPos - m_pFile->GetPointer(0);
+	Ctx.dwPos1 = (int)(pPos - m_pFile->GetPointer(0));
 
 	//	Done
